通过的测试点微服务项目qa

+++ tensquare_user/src/main/java/com/tensquare/user/service/UserService.java
@@ -3,19 +3,23 @@ package com.tensquare.user.service;
 import com.tensquare.user.dao.UserDao;
 import com.tensquare.user.pojo.User;
 import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import util.IdWorker;
+import util.JwtUtil;
 
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -41,6 +45,29 @@ public class UserService {
 
     @Autowired
     private RabbitTemplate rabbitTemplate;
+
+    @Autowired
+    private BCryptPasswordEncoder encoder;
+
+    @Autowired
+    private HttpServletRequest httpServletRequest;
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    @Autowired
+    private HttpServletRequest request;
+
+    public User login(String mobile, String password) {
+        //先根据用户名查询对象
+        User userLogin = userDao.findByMobile(mobile);
+        //然后拿数据库中的密码和用户输入的密码匹配是否相同
+        if (userLogin != null && encoder.matches(password, userLogin.getPassword())) {
+            return userLogin;
+        }
+        //登录失败
+        return null;
+    }
 
     public void sendsms(String mobile) {
         //生成六位随机数字
@@ -106,6 +133,8 @@ public class UserService {
      * @param user
      */
     public void add(User user) {
+        //用户的加密
+        user.setPassword(encoder.encode(user.getPassword()));
         user.setId(idWorker.nextId() + "");
         userDao.save(user);
     }
@@ -120,11 +149,15 @@ public class UserService {
     }
 
     /**
-     * 删除
+     * 删除 必须是删除admin这个角才能删除
      *
      * @param id
      */
     public void deleteById(String id) {
+        String token = (String) request.getAttribute("claims_admin");
+        if (StringUtils.isEmpty(token)) {
+            throw new RuntimeException("权限不足");
+        }
         userDao.deleteById(id);
     }
 
@@ -135,9 +168,7 @@ public class UserService {
      * @return
      */
     private Specification<User> createSpecification(Map searchMap) {
-
         return new Specification<User>() {
-
             @Override
             public Predicate toPredicate(Root<User> root, CriteriaQuery<?> query, CriteriaBuilder cb) {
                 List<Predicate> predicateList = new ArrayList<Predicate>();
@@ -177,12 +208,8 @@ public class UserService {
                 if (searchMap.get("personality") != null && !"".equals(searchMap.get("personality"))) {
                     predicateList.add(cb.like(root.get("personality").as(String.class), "%" + (String) searchMap.get("personality") + "%"));
                 }
-
                 return cb.and(predicateList.toArray(new Predicate[predicateList.size()]));
-
             }
         };
-
     }
-
 }
